"Refactor error check in unarchiver extract function
Changed the error check comparison in the extract_archive function within unarchiver.c file. Comparing result with (-1u) instead of (uint32_t)-1 improves type checking. This insures consistency across the codebase and promotes code reliability by reducing possibility of typecasting errors."

+++ src/tp2/unarchiver.c
@@ -162,7 +162,7 @@ int main(int argc, char **argv)
     }
 
     uint32_t result = extract_archive(argv[1]);
-    if (result == (uint32_t)-1)
+    if (result == (-1u))
     {
         perror("Error extracting archive");
         return 1;
